+++ apps/api/src/notification/application/use-cases/get-unread-count/get-unread-count.use-case.ts
@@ -5,9 +5,9 @@ import {
 	type NotificationCachePort,
 } from "../../ports/notification-cache.port";
 import {
-	NOTIFICATION_REPOSITORY,
-	type NotificationRepositoryPort,
-} from "../../ports/notification.repository.port";
+	NOTIFICATION_INBOX_READER,
+	type NotificationInboxReaderPort,
+} from "../../ports/notification-inbox.reader.port";
 
 /**
  * 읽지 않은 알림 수 조회 유스케이스 (2분 캐시).
@@ -15,15 +15,15 @@ import {
 @Injectable()
 export class GetUnreadCountUseCase {
 	constructor(
-		@Inject(NOTIFICATION_REPOSITORY)
-		private readonly notificationRepository: NotificationRepositoryPort,
+		@Inject(NOTIFICATION_INBOX_READER)
+		private readonly notificationInboxReader: NotificationInboxReaderPort,
 		@Inject(NOTIFICATION_CACHE)
 		private readonly cache: NotificationCachePort,
 	) {}
 
 	async execute(userId: string): Promise<number> {
 		return this.cache.wrapUnreadCount(userId, () =>
-			this.notificationRepository.countUnread(userId),
+			this.notificationInboxReader.countUnread(userId),
 		);
 	}
 }
